requirements.txt
jupyter
matplotlib
sympy
pandas
numpy
seaborn
scipy
pyyaml
sklearn